refactor: 重构

+++ src/main/java/util/DebugUtil.java
@@ -1,10 +1,13 @@
 package util;
 
+/**
+ * @author weloe
+ */
 public class DebugUtil {
 
     int level = 1;
 
-    public static void log(String msg,Object ... objects){
+    public static void logf(String msg, Object... objects) {
         System.out.printf(msg + "\n", objects);
     }
